Initialise GL state in init() so widgets like Text will work.
The Text widget (or FTGL, for that matter) has problems when calling FTGL methods before OpenGL
matrix setup has taken place. This resulted in messed-up glyphs. The problem was worked around
by updating Texts via preRenderUpdate(). However, this approach creates problems if we want to
introduce auto-sized Labels. At the moment the Label's text is changed, the internal Text widget
should calculate its new size and resize its containing Label accordingly. Due to the work-around,
though, that size update has to wait until preRenderUpdate() is called. This is not a problem
generally, but if something depends on the size of that Label (including listeners) there might
occur problems or inconsistencies.

In order to have a direct "reaction", the solution is to initialise the OpenGL state earlier
(outside of the main loop, in init()) so that FTGL and other similarly sensitive code will work
properly.